app/src/main/java/com/codesoom/assignment/products/application/ProductService.java
package com.codesoom.assignment.products.application;

import com.codesoom.assignment.products.domain.Product;
import com.codesoom.assignment.products.domain.ProductRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductService {
    private final ProductRepository productRepository;

    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public List<Product> getTasks() {
        return productRepository.findAll();
    }
}
